+++ src/Server/OperationParams.php
@@ -5,21 +5,25 @@ declare(strict_types=1);
 namespace GraphQL\Server;
 
 use function array_change_key_case;
+use function array_walk;
 use function is_string;
 use function json_decode;
 use function json_last_error;
-use function strlen;
 
 use const CASE_LOWER;
 use const JSON_ERROR_NONE;
 
 /**
- * Structure representing parsed HTTP parameters for GraphQL operation
+ * Structure representing parsed HTTP parameters for GraphQL operation.
+ *
+ * The properties in this class are not strictly typed, as this class
+ * is only meant to serve as an intermediary representation which is
+ * not yet validated.
  */
 class OperationParams
 {
     /**
-     * Id of the query (when using persistent queries).
+     * Id of the query (when using persisted queries).
      *
      * Valid aliases (case-insensitive):
      * - id
@@ -27,38 +31,56 @@ class OperationParams
      * - documentId
      *
      * @api
-     * @var string|null
+     * @var mixed should be string|null
      */
     public $queryId;
 
     /**
+     * A document containing GraphQL operations and fragments to execute.
+     *
      * @api
-     * @var string|null
+     * @var mixed should be string|null
      */
     public $query;
 
     /**
+     * The name of the operation in the document to execute.
+     *
      * @api
-     * @var string
+     * @var mixed should be string|null
      */
     public $operation;
 
     /**
+     * Values for any variables defined by the operation.
+     *
      * @api
-     * @var array<string, mixed>|null
+     * @var mixed should be array<string, mixed>
      */
     public $variables;
 
     /**
+     * Reserved for implementors to extend the protocol however they see fit.
+     *
      * @api
-     * @var array<string, mixed>|null
+     * @var mixed should be array<string, mixed>
      */
     public $extensions;
 
-    /** @var array<string, mixed> */
-    private $originalInput;
+    /**
+     * Executed in read-only context (e.g. via HTTP GET request)?
+     *
+     * @api
+     */
+    public bool $readOnly;
 
-    private bool $readOnly;
+    /**
+     * The raw params used to construct this instance.
+     *
+     * @api
+     * @var array<string, mixed>
+     */
+    public array $originalInput;
 
     /**
      * Creates an instance from given array
@@ -84,33 +106,27 @@ class OperationParams
             'extensions' => null,
         ];
 
-        if ($params['variables'] === '') {
-            $params['variables'] = null;
-        }
-
-        // Some parameters could be provided as serialized JSON.
-        foreach (['extensions', 'variables'] as $param) {
-            if (! is_string($params[$param])) {
-                continue;
+        array_walk($params, static function (&$value): void {
+            if ($value !== '') {
+                return;
             }
 
-            $tmp = json_decode($params[$param], true);
-            if (json_last_error() !== JSON_ERROR_NONE) {
-                continue;
-            }
-
-            $params[$param] = $tmp;
-        }
+            $value = null;
+        });
 
         $instance->query      = $params['query'];
         $instance->queryId    = $params['queryid'] ?? $params['documentid'] ?? $params['id'];
         $instance->operation  = $params['operationname'];
-        $instance->variables  = $params['variables'];
-        $instance->extensions = $params['extensions'];
+        $instance->variables  = static::decodeIfJSON($params['variables']);
+        $instance->extensions = static::decodeIfJSON($params['extensions']);
         $instance->readOnly   = $readonly;
 
-        // Apollo server/client compatibility: look for the queryid in extensions
-        if (isset($instance->extensions['persistedQuery']['sha256Hash']) && strlen($instance->query ?? '') === 0 && strlen($instance->queryId ?? '') === 0) {
+        // Apollo server/client compatibility
+        if (
+            isset($instance->extensions['persistedQuery']['sha256Hash'])
+            && $instance->query === null
+            && $instance->queryId === null
+        ) {
             $instance->queryId = $instance->extensions['persistedQuery']['sha256Hash'];
         }
 
@@ -118,9 +134,9 @@ class OperationParams
     }
 
     /**
-     * @return mixed
+     * @deprecated
      *
-     * @api
+     * @return mixed
      */
     public function getOriginalInput(string $key)
     {
@@ -128,13 +144,31 @@ class OperationParams
     }
 
     /**
-     * Indicates that operation is executed in read-only context
-     * (e.g. via HTTP GET request)
-     *
-     * @api
+     * @deprecated
      */
     public function isReadOnly(): bool
     {
         return $this->readOnly;
+    }
+
+    /**
+     * Decodes the value if it is JSON, otherwise returns it unchanged.
+     *
+     * @param mixed $value
+     *
+     * @return mixed
+     */
+    protected static function decodeIfJSON($value)
+    {
+        if (! is_string($value)) {
+            return $value;
+        }
+
+        $decoded = json_decode($value, true);
+        if (json_last_error() === JSON_ERROR_NONE) {
+            return $decoded;
+        }
+
+        return $value;
     }
 }